requirements.txt
av==10.0.0
colorama==0.4.6
controlnet_aux==0.0.7
diffusers==0.21.4
fastapi==0.103.2
fire==0.5.0
llama_cpp_python==0.1.84
nest_asyncio==1.5.6
numpy==1.24.0
Pillow==10.0.1
pydantic==1.10.9
pyngrok==7.0.0
requests==2.28.0
safetensors==0.3.1
tomesd==0.1.3
torch==2.0.1
transformers==4.30.1
uvicorn==0.23.2
